requirements.txt
# --- GUI & Core ---
PyQt6>=6.4.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=9.5.0
pyyaml>=6.0

# --- AI & Deep Learning ---
# หมายเหตุ: Torch แนะนำให้ลงแยกตาม version CUDA ของเครื่อง
# pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0
torchvision>=0.15.0

# --- Object Detection (YOLO) ---
ultralytics>=8.0.0

# --- Classifier Backbone ---
timm>=0.9.0

# --- Utilities ---
tqdm